floating point number fix

+++ model.py
@@ -92,7 +92,7 @@ def constructModel() -> AgentModel:
     model.update_parameters(
         {
             "num_agents": 3,
-            "proportion_marginalized": 1 / 6,
+            "proportion_marginalized": float(1 / 6),
             "num_pulls": 1,
             "objective_b": 0.51,
         }
